+++ packages/ozone/src/daemon/event-pusher.ts
@@ -211,7 +211,7 @@ export class EventPusher {
             ? { confirmedAt: new Date() }
             : {
                 lastAttempted: new Date(),
-                attempts: evt.attempts ?? 0 + 1,
+                attempts: (evt.attempts ?? 0) + 1,
               },
         )
         .where('subjectDid', '=', evt.subjectDid)
@@ -250,7 +250,7 @@ export class EventPusher {
             ? { confirmedAt: new Date() }
             : {
                 lastAttempted: new Date(),
-                attempts: evt.attempts ?? 0 + 1,
+                attempts: (evt.attempts ?? 0) + 1,
               },
         )
         .where('subjectUri', '=', evt.subjectUri)
@@ -331,7 +331,7 @@ export class EventPusher {
           lastAttempted: null,
         }),
       )
-      .returning('id')
+      .returning(['id', 'subjectDid', 'subjectUri', 'subjectBlobCid', 'eventType'])
       .execute()
   }
 }