+++ packages/lex-cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atproto/lex-cli
 
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies [[`0c815b964`](https://github.com/bluesky-social/atproto/commit/0c815b964c030aa0f277c40bf9786f130dc320f4)]:
+  - @atproto/syntax@0.2.0
+  - @atproto/lexicon@0.3.2
+
 ## 0.3.0
 
 ### Minor Changes