+++ packages/xrpc-server/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @atproto/xrpc-server
 
+## 0.4.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @atproto/lexicon@0.3.2
+
 ## 0.4.2
 
 ### Patch Changes